+++ src/components/modules/escrow/ui/dialogs/hooks/fund-escrow-dialog.hook.ts
@@ -6,7 +6,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 import { formSchema } from "../../../schema/fund-escrow.schema";
-import { fundEscrow } from "@/components/modules/escrow/services/fundEscrow";
+import { fundEscrow } from "@/components/modules/escrow/services/fund-escrow.service";
 import {
   useGlobalAuthenticationStore,
   useGlobalBoundedStore,
