+++ src/core/store/data/slices/escrows.slice.ts
@@ -6,7 +6,7 @@ import {
   getAllEscrowsByUser,
   updateEscrow,
 } from "@/components/modules/escrow/server/escrow.firebase";
-import { getBalance } from "@/components/modules/escrow/services/getBalance";
+import { getBalance } from "@/components/modules/escrow/services/get-balance.service";
 
 const ESCROW_ACTIONS = {
   SET_ESCROWS: "escrows/set",
